Refactor code into Fixes/Patches/Features modules
Split the monolithic dllmain.cpp into category modules and run them from a
config -> apply table:

- New folders: Fixes/, Patches/, Features/ (source + matching namespaces,
  e.g. Fixes::Graphics::VSync, Patches::System::Telemetry).
- common.h replaces TypeDefs.h; flat includes, no '..\' relative paths.
- Remove the old monolith files (InputHooks, WindowHooks, framework.h).
- Config grouped into Config::Fixes / Patches / Features namespaces; ini keys
  keep their original names + help text. New UseHighPrecisionTimer flag for the
  high-precision timer (physics) fix, default on.
- dllmain drives a ModuleEntry table (enabled() && apply()).
- vcxproj/filters updated to the new file layout; $(ProjectDir) added to the
  include path so common.h resolves.
