+++ DeadSpaceFixes/Fixes/Fixes.h
@@ -1,0 +1,45 @@
+#pragma once
+
+// Bug-fix patches: correct broken or incorrect engine behaviour.
+// Each fix is a self-contained .cpp file (listed in the project) exposing a
+// single `Apply(HMODULE)` entry point declared here.
+
+#include <Windows.h>
+
+namespace Fixes {
+
+	namespace Physics {
+		namespace Timer {
+			void Apply(HMODULE hExe);
+		}
+	}
+
+	namespace Graphics {
+		namespace VSync {
+			void ApplyStartup();
+			void Apply(HMODULE hExe);
+		}
+		namespace SubtitleScale {
+			void Apply(HMODULE hExe);
+		}
+	}
+
+	namespace UI {
+		namespace LoadingScreen {
+			void Apply(HMODULE hExe);
+		}
+	}
+
+	namespace Input {
+		namespace LegacyDirectInput {
+			void Apply(HMODULE hExe);
+		}
+	}
+
+	namespace Save {
+		namespace SafeStringHandling {
+			void Apply(HMODULE hExe);
+		}
+	}
+
+}
