--- DeadSpaceFixes/InputHooks.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <Windows.h>
-#include "TypeDefs.h"
-
-namespace Input {
-	void InitialiseInputHooks();
-}
-bool hkIsGamePad();
-
-extern "C" DWORD WINAPI XInputGetState(DWORD dwUserIndex, XINPUT_STATE* pState);
-extern "C" DWORD WINAPI XInputSetState(DWORD dwUserIndex, XINPUT_VIBRATION* pVibration);
-extern "C" DWORD WINAPI XInputGetCapabilities(DWORD dwUserIndex, DWORD dwFlags, XINPUT_CAPABILITIES* pCapabilities);
